+++ tests/Fixtures/ConsoleSample/Console/AbstractSampleCommand.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Dem1Off\LaravelModular\Tests\Fixtures\ConsoleSample\Console;
+
+use Illuminate\Console\Command;
+
+abstract class AbstractSampleCommand extends Command {}